requirements.txt
google-cloud-storage
transformers
torch
pandas
numpy
scikit-learn
python-dotenv
-e .